chore: separate coroutines code to coroutines module
